more function more loop

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -5,7 +5,7 @@
  */
 int main(void)
 {
-	long int n, div = 2, ans = 0, maxFact;
+	long long int n, div = 2, ans = 0, maxFact;
 
 	n = 612852475143;
 	while (n != 0)
